[WIP] Trying to remove custom qt patch dependency
Hi, I really liked the idea of the project. I thought maybe these
changes could achieve what the patch is trying to. I'm not completely
sure though, there appears one more permission issue that the patch is
addressing which this commit is not.

Please have a look at the changes, I'm completely new to QtBLE programming
so there might be some mess I made. But if not, if I'm on the correct
track, I'd really like to proceed with the changes and address
everything that the patch addreseses.